test for parse function which check if item is null on invalid elements
Test which checks if a null item is returned when there are no elements. Since the code implementation of this test was wrongly implemented on previous test, it was removed for the purpose of this test from the function and will be re-added later, to prove test logic works.
Second test checks if there are elements in the function but they do not fit one of the three type of items. Verifies if no parseBook,parseMovie or parseMusic is called.